fix(ci): move beads SDK dep to integration test file to fix build (gt-9wg)
The beads SDK (CGo/Dolt) can't be imported in the helpers file which
has no build tag. Move ensureCrossRigDep to scheduler_integration_test.go
which has //go:build integration.
