+++ internal/cmd/scheduler_integration_test.go
@@ -13,6 +13,7 @@
 package cmd
 
 import (
+	"context"
 	"database/sql"
 	"encoding/json"
 	"fmt"
@@ -24,6 +25,7 @@ import (
 	"testing"
 	"time"
 
+	beadsdk "github.com/steveyegge/beads"
 	"github.com/steveyegge/gastown/internal/beads"
 	"github.com/steveyegge/gastown/internal/config"
 	"github.com/steveyegge/gastown/internal/scheduler/capacity"
@@ -233,6 +235,28 @@ func findSlingContext(t *testing.T, hqPath, workBeadID string) *capacity.SlingCo
 func hasSlingContext(t *testing.T, hqPath, workBeadID string) bool {
 	t.Helper()
 	return findSlingContext(t, hqPath, workBeadID) != nil
+}
+
+// ensureCrossRigDep creates a cross-rig dependency via the beads Go SDK,
+// writing directly to the embedded store. Used when bd v1.0.0 can't create
+// cross-rig deps via the CLI (which validates target existence).
+func ensureCrossRigDep(t *testing.T, from, to, depType, dir string) {
+	t.Helper()
+	ctx := context.Background()
+	beadsDir := filepath.Join(dir, ".beads")
+	store, err := beadsdk.OpenFromConfig(ctx, beadsDir)
+	if err != nil {
+		t.Fatalf("ensureCrossRigDep: open store at %s: %v", beadsDir, err)
+	}
+	defer store.Close()
+	dep := &beadsdk.Dependency{
+		IssueID:     from,
+		DependsOnID: to,
+		Type:        beadsdk.DependencyType(depType),
+	}
+	if err := store.AddDependency(ctx, dep, "test"); err != nil {
+		t.Fatalf("ensureCrossRigDep: add dep %s→%s: %v", from, to, err)
+	}
 }
 
 // --------------------------------------------------------------------------
